Deployment/Backend/requirements.txt
waitress
pandas
numpy
scikit-learn
flask
tensorflow-cpu
